Add MetricsProcessor class
This commit takes the methods merge_metrics and condense_metrics out of utils.py
and puts them into the MetricsProcessor class.

The callers have been updated to instantiate this class and use the methods
provided by it.

The unit tests have been moved to their own module. The tests needed some
minor updates to run but otherwise have remained unchanged.
