+++ openshift_metrics/utils.py
@@ -207,123 +207,6 @@ def get_service_unit(cpu_count, memory_count, gpu_count, gpu_type, gpu_resource)
     return su_type, su_count, determining_resource
 
 
-def merge_metrics(metric_name, metric_list, output_dict):
-    """
-    Merge metrics by pod but since pod names aren't guaranteed to be unique across
-    namespaces, we combine the namespace and podname together when generating the
-    output dictionary so it contains all pods.
-    """
-
-    for metric in metric_list:
-        pod = metric["metric"]["pod"]
-        namespace = metric["metric"]["namespace"]
-        node = metric["metric"].get("node")
-
-        gpu_type = None
-        gpu_resource = None
-        node_model = None
-
-        unique_name = namespace + "+" + pod
-        if unique_name not in output_dict:
-            output_dict[unique_name] = {"namespace": namespace, "metrics": {}}
-
-        if metric_name == "gpu_request":
-            gpu_type = metric["metric"].get("label_nvidia_com_gpu_product", GPU_UNKNOWN_TYPE)
-            gpu_resource = metric["metric"].get("resource")
-            node_model = metric["metric"].get("label_nvidia_com_gpu_machine")
-
-        for value in metric["values"]:
-            epoch_time = value[0]
-            if epoch_time not in output_dict[unique_name]["metrics"]:
-                output_dict[unique_name]["metrics"][epoch_time] = {}
-            output_dict[unique_name]["metrics"][epoch_time][metric_name] = value[1]
-            if gpu_type:
-                output_dict[unique_name]["metrics"][epoch_time]['gpu_type'] = gpu_type
-            if gpu_resource:
-                output_dict[unique_name]["metrics"][epoch_time]['gpu_resource'] = gpu_resource
-            if node_model:
-                output_dict[unique_name]["metrics"][epoch_time]['node_model'] = node_model
-            if node:
-                output_dict[unique_name]["metrics"][epoch_time]['node'] = node
-
-    return output_dict
-
-
-def condense_metrics(input_metrics_dict, metrics_to_check):
-    """
-    Checks if the value of metrics is the same, and removes redundant
-    metrics while updating the duration. If there's a gap in the reported
-    metrics then don't count that as part of duration.
-
-    Here's a sample input dictionary in which I have separated missing metrics
-    or different metrics by empty lines.
-
-    {'naved-test+test-pod': {'gpu_type': 'No GPU',
-                         'metrics': {1711741500: {'cpu_request': '1',
-                                                  'memory_request': '3221225472'},
-                                     1711742400: {'cpu_request': '1',
-                                                  'memory_request': '3221225472'},
-                                     1711743300: {'cpu_request': '1',
-                                                  'memory_request': '3221225472'},
-                                     1711744200: {'cpu_request': '1',
-                                                  'memory_request': '3221225472'},
-
-                                     1711746000: {'cpu_request': '1',
-                                                  'memory_request': '3221225472'},
-
-                                     1711746900: {'cpu_request': '1',
-                                                  'memory_request': '4294967296'},
-                                     1711747800: {'cpu_request': '1',
-                                                  'memory_request': '4294967296'},
-                                     1711748700: {'cpu_request': '1',
-                                                  'memory_request': '4294967296'},
-
-                                     1711765800: {'cpu_request': '1',
-                                                  'memory_request': '4294967296'}},
-                         'namespace': 'naved-test'}}
-    """
-    interval = STEP_MIN * 60
-    condensed_dict = {}
-    for pod, pod_dict in input_metrics_dict.items():
-        metrics_dict = pod_dict["metrics"]
-        new_metrics_dict = {}
-        epoch_times_list = sorted(metrics_dict.keys())
-
-        start_epoch_time = epoch_times_list[0]
-
-        start_metric_dict = metrics_dict[start_epoch_time].copy()
-
-        for i in range(len(epoch_times_list)):
-            epoch_time = epoch_times_list[i]
-            same_metrics = True
-            continuous_metrics = True
-            for metric in metrics_to_check:
-                if metrics_dict[start_epoch_time].get(metric, 0) != metrics_dict[epoch_time].get(metric, 0):  # fmt: skip
-                    same_metrics = False
-
-            if i !=0 and epoch_time - epoch_times_list[i-1]> interval:
-                # i.e. if the difference between 2 consecutive timestamps
-                # is more than the expected frequency then the pod was stopped
-                continuous_metrics = False
-
-            if not same_metrics or not continuous_metrics:
-                duration = epoch_times_list[i-1] - start_epoch_time + interval
-                start_metric_dict["duration"] = duration
-                new_metrics_dict[start_epoch_time] = start_metric_dict
-                start_epoch_time = epoch_time
-                start_metric_dict = metrics_dict[start_epoch_time].copy()
-
-        duration = epoch_time - start_epoch_time + interval
-        start_metric_dict["duration"] = duration
-        new_metrics_dict[start_epoch_time] = start_metric_dict
-
-        new_pod_dict = pod_dict.copy()
-        new_pod_dict["metrics"] = new_metrics_dict
-        condensed_dict[pod] = new_pod_dict
-
-    return condensed_dict
-
-
 def csv_writer(rows, file_name):
     """Writes rows as csv to file_name"""
     print(f"Writing csv to {file_name}")
